requirements.txt
alembic==1.0.5
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
codeclimate-test-reporter==0.2.3
coverage==4.3.1
Flask==0.12.3
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.15.1
marshmallow-sqlalchemy==0.13.2
more-itertools==4.3.0
parso==0.3.1
pg8000==1.12.3
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
pytest==4.0.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.0.0
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.24.1
Werkzeug==0.14.1
